Accept parameters for setDadaOpt
params.dadaOpt can be specified as a dictionary in a config
or as --dadaOpt.XXX VALUE
( providing nested scope params from command line requires Nextflow 20.11.0)

The params are passed to scripts where learnErrors() or dada() are called.

A default dadaOpt = [] removes a startup warning. Deprecate a similar dadaParams in pacbio.nf